requirements.txt
anyio==3.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
gTTS==2.3.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
pycparser==2.21
pydub==0.25.1
python-telegram-bot==20.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
SpeechRecognition==3.9.0
urllib3==1.26.14

TTS~=0.10.2
python-dotenv~=0.21.1
aiosqlite~=0.18.0
pyttsx3~=2.90